刪除的功能 2

+++ address-book/list.php
@@ -73,7 +73,10 @@ $rows = $pdo->query($sql)->fetchAll();
                 <?php foreach($rows as $r): ?>
                     <tr>
                         <td>
-                            <a href="delete.php?sid=<?= $r['sid'] ?>">
+                            <!--
+                            <a href="delete.php?sid=<?= $r['sid'] ?>" onclick="return confirm('確定要刪除這筆資料嗎?')">
+                            -->
+                            <a href="javascript: delete_it(<?= $r['sid'] ?>)">
                                 <i class="fas fa-trash-alt"></i>
                             </a>
                         </td>
@@ -101,6 +104,15 @@ $rows = $pdo->query($sql)->fetchAll();
 </div>
 
 <?php include __DIR__. '/parts/__scripts.php' ?>
+<script>
+    function delete_it(sid){
+        if(confirm(`確定要刪除編號為 ${sid} 的資料嗎?`)){
+            location.href = `delete.php?sid=${sid}`;
+        }
+    }
+
+
+</script>
 <?php include __DIR__. '/parts/__html_foot.php' ?>
